refactor(port): update BeneficiaryAccountRepositoryPort with Long ID
Función: Puerto de salida para la persistencia de cuentas de beneficiarios.
Cambios: Se actualizó la firma del método 'softDelete' para aceptar un identificador de tipo Long.

+++ src/main/java/org/walrex/application/port/output/BeneficiaryAccountRepositoryPort.java
@@ -6,5 +6,5 @@ import org.walrex.domain.model.BeneficiaryAccount;
 public interface BeneficiaryAccountRepositoryPort {
     Uni<BeneficiaryAccount> save(BeneficiaryAccount beneficiaryAccount);
     Uni<BeneficiaryAccount> update(BeneficiaryAccount beneficiaryAccount);
-    Uni<Boolean> softDelete(Integer id);
+    Uni<Boolean> softDelete(Long id);
 }
